requirements.txt
# Core app
streamlit>=1.25.0
streamlit-drawable-canvas>=0.9.2

# Image handling
Pillow>=9.0.0
opencv-python>=4.5.5.64
scikit-image>=0.19.3

# Data handling
numpy>=1.22.0

# Visualization (for scoring feedback)
matplotlib>=3.5.1

# Optional: for development and debugging
ipython
